CHROMIUM: net: Make "uidrange 0-n" match kernel packets in containers
Kernel-generated packets, such as ICMP echo replies, are implicitly
treated as KUID 0 packets if there is a uidmatch IP routing rule in place.
But inside a userns, it usually is not possible to create a uidmatch
rule that covers KUID 0, because UID 0 is typically mapped to a higher
value.  Add a special case to the routing code that carries over the old
behavior to the container scenario: if a uidrange starts at UID 0,
always match KUID 0 packets too.

BUG=b:27248583
TEST=`run cts --package android.net.hostsidenetwork --disable-reboot`

Change-Id: Id9acb0fb60565ffc234d22fb1ad9f00c9e129318
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/361973
(cherry picked from commit 48bf835f1d81876734db6cb2a1afc6db4cd28a07)
Reviewed-on: https://chromium-review.googlesource.com/362777