UPSTREAM: [media] cec: zero unused msg part after msg->len
Ensure that the unused part of the msg array is zeroed. This is
required by CEC 2.0 when receiving shorter messages than expected.
In that case the remaining bytes are assumed to be 0.

There are no such CEC messages yet, but it is required to be future proof.

And since we're doing it for received messages, do it for transmitted
messages as well. It's unambiguous this way.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit 045344c3c1478f8d03a1a589d645fe10c3a23f77)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I030d3a4037a6d9cdaec5d42728522f9dc126b1a4
Reviewed-on: https://chromium-review.googlesource.com/1021266
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>