REDO: CHROMIUM: HACK: vop0 (big) pixel clock owns VPLL, nobody else uses
There are a few possible problems with the 'normal' VOP clk tree:
1. At the whim of firmware, it's possible that VPLL will be initialized
to something that is a multiple of our chosen pixel clock (266666 MHz).
At the whim of common clock framework it's possible that VPLL will be
chosen as the parent clock of the little VOP and be used by eDP. Now
when we plug into type C then we'll want to change VPLL. Common clock
framework is not currently smart enough to know that it should reparent
the little VOP to CPLL / GPLL in this case. Instead the eDP pixel clock
will be unexpectedly changed. Note that there are some "whims" there,
but generally I don't like relying on un-guaranteed behavior.

2. Even if little vop gets cpll / gpll (YAY), the common clock framework
could (on a whim) decide to also point the big VOP to cpll / gpll, maybe
because you plug into a external display that happens to have pixel
clock that can be made from cpll / gpll. Then you might change your
display mode. Rather than reparent the big VOP back to vpll, the CCF
might (on a whim) decide to change the rate of CPLL / GPLL, which would
be a terrible idea. To summarize: the CCF has too much "automatic"
behavior if you've got both CLK_SET_RATE_PARENT and you allow
auto-remuxing (the default).  It's going to bite us unless we do
something to make sure that only the expected parents are used.

Mark dummy for all the PLL which we don't want to assign to someone.
That will give VPLL to big VOP. Little VOP will dynamically switch
between CPLL and GPLL depending on the requested pixel clock.

BUG=chrome-os-partner:52872
TEST=Check DP and eDP display

Change-Id: Ic0b244845d70a6bbbc069c28b2a2263ff9dd465f
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/362820
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/384397
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Trybot-Ready: Douglas Anderson <dianders@chromium.org>