ANDROID: FIX: sched/tune: update usage of boosted task utilisation on CPU selection
A boosted task needs to be scheduled on a CPU which can grant a minimum
capacity which is higher than its utilization.
However, a task can be allocated on a CPU which already provides an utilization
which is higher than the task boosted utilization itself.
Moreover, with the previous approach a task 100% boosted is not fitting any
CPU.

This patch makes use of the boosted task utilization just as a threashold
which defines the minimum capacity should be available on a CPU to host that
task.

BUG=chrome-os-partner:58667
TEST=build/boot on kevin

Change-Id: I5b0f62925889f84df32b0f85c1aaaf0c150c94bb
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit 28e8cb961c2f71a89e391335a9304c3dd8b38d8
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/420055
(cherry picked from commit c373636ba27e64f2006d1a757c18c4b488efa6c1)
Reviewed-on: https://chromium-review.googlesource.com/424042