CHROMIUM: iwl7000: mvm: add initiator XVT capabilities
Add initiator flags for the range request.
The flags define the measurement process. for
example: which algorithm type to run and what are the
calibration values needs to be factored in.
Add debugfs control to the initiator flags. This shall
allow a user to override the default flag values and
control the measurement process.
Add range notification flags to enable to report
internal and partial algorithm products. for
example: t1t4_responder shall contain the result of the
algorithm on the responder side as was reported to the
initiator in the FTM action frame.

Signed-off-by: Eyal Zolotov <eyal.zolotov@intel.com>
iwl7000-tree: 384825e45434e4e40fc76e13c3f5e64c416751c5