UPSTREAM: drm/i915: Refactor golden render state emission to unconfuse gcc
GCC was inlining the init and setup functions, but was getting itself
confused into thinking that variables could be used uninitialised. If we
do the inline for gcc, it is happy! As a bonus we shrink the code.

v2: A couple of minor tweaks from Joonas

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1469432687-22756-29-git-send-email-chris@chris-wilson.co.uk
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1470174640-18242-20-git-send-email-chris@chris-wilson.co.uk
(cherry picked from commit 15d21db87239840ae8248cf26e46391edc13cfe3)
from drm-intel-next-2016-08-08
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

Change-Id: Iba8e36a8e65bdced7a2ee23afc28b574c66716ae