UPSTREAM: drm/dp_helper: Print first error received on failure in drm_dp_dpcd_access()
Since we always retry in drm_dp_dpcd_access() regardless of the error,
we're going to make a lot of noise if the aux->transfer function prints
it's own errors (as is the case with radeon). If we can print the error
code here, this reduces the need for drivers to do this. So instead of
having to print "dp_aux_ch timed out" over 32 times we can just print
once.

BUG=chrome-os-partner:60130,chrome-os-partner:60965
TEST=build and boot samus

Signed-off-by: Lyude <cpaul@redhat.com>
Reviewed-by: Christian Knig <christian.koenig@amd.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Link: http://patchwork.freedesktop.org/patch/msgid/1470443443-27252-2-git-send-email-cpaul@redhat.com
(cherry picked from commit 29f21e0491819d05d28b4a1f7a5b7e68f802c60b)
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Change-Id: Ib00551f769d2775c4ec465c78360f0e0674c294d
Reviewed-on: https://chromium-review.googlesource.com/433195
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>