UPSTREAM: staging/android: add Doc for SW_SYNC ioctl interface
This interface is hidden from kernel headers and it is intended for use
only for testing. So testers would have to add the ioctl information
internally. This is to prevent misuse of this feature.

v2: take in Eric suggestions for the Documentation

v3: really take in Eric suggestions

BUG=chromium:911825
TEST=boot caroline, chell, run login_LoginSuccess, graphics_SanAngeles, cheets_StartAndroid

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.co.uk>
Reviewed-by: Eric Engestrom <eric.engestrom@imgtec.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit fc0c9a03b57efb769a74fa8b33d68d90153b5950)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I330b7b4d99f6d20e8bc61146f9b40cb80b0192bd