UPSTREAM: Bluetooth: Fix bt_sock_recvmsg when MSG_TRUNC is not set
Commit b5f34f9420b50c9b5876b9a2b68e96be6d629054 attempt to introduce
proper handling for MSG_TRUNC but recv and variants should still work
as read if no flag is passed, but because the code may set MSG_TRUNC to
msg->msg_flags that shall not be used as it may cause it to be behave as
if MSG_TRUNC is always, so instead of using it this changes the code to
use the flags parameter which shall contain the original flags.

BUG=chromium:704196
TEST=Deploy v5.44 BlueZ system daemon with v4.9 BlueZ kernel on Oak and
perform automated tests (suite:bluetooth) and manual tests against mouses,
keyboards and headphones.

Change-Id: I6532300e6cdb1c7ae22104418dcdd1a78b8e9a77
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Miao-chen Chou <mcchou@chromium.org>
(cherry picked from commit 90a56f72edb088c678083c32d05936c7c8d9a948)
Reviewed-on: https://chromium-review.googlesource.com/474468
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>